Add maximized trading feature
Added a new 'maximize' feature for uranium and coal trading. When in 'maximize' mode, instead of performing the minimum number of trades required to fill the resource stockpile, we instead perform the maximum number of trades possible, converting existing uranium or coal into thorium or steel first to make room.

This feature is controlled by a pair of checkboxes I added to the script main menu under the AutoTrade button. In order to make room for this, I moved the menu up 200 pixels. While I was at it, I moved the AutoBuild building select up 100 pixels to make everything visible.