Adjust Leviathan trading conditions
Trading with the Leviathans causes the remaining duration of their visit to be reduced to a certain cap (the exact value depends on their energy level). So, to maximize the number of trades possible per visit, instead of trading as early and as often as possible, we now only trade if the remaining duration is already under that cap, _or_ if our unobtanium stockpile is full.